Clean unwrapping kpts notebook
